Last updates from Master
Update ReadMe
+
Yosinda = update
Assign "HELP" key to F11 key

In the original, the PC88's "HELP" key is assigned to the F12 key on the keyboard.
However, the F12 key is already used for the MiSTer OSD menu, so it cannot be used in the core.

Change the assignment of "HELP" to the F11 key.